refactor!: rename persistence namespace to unit of work
